feat: city delete by id returning exception when not found

+++ src/main/java/com/devsuperior/demo/services/CityService.java
@@ -3,6 +3,7 @@ package com.devsuperior.demo.services;
 import com.devsuperior.demo.dto.CityDTO;
 import com.devsuperior.demo.entities.City;
 import com.devsuperior.demo.repositories.CityRepository;
+import com.devsuperior.demo.services.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -37,6 +38,10 @@ public class CityService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new NotFoundException("City not founded");
+        }
+
         repository.deleteById(id);
     }
 }
